Disable buttons on event page, move to it's own page for easier importing
